+++ tests/test_hybrid.py
@@ -12,7 +12,7 @@ from micom.workflows import (
     tradeoff,
     minimal_media,
     complete_community_medium,
-    GrowthResults
+    GrowthResults,
 )
 from micom.qiime_formats import load_qiime_medium
 from micom.solution import CommunitySolution, OptimizationError
@@ -20,8 +20,7 @@ import pytest
 from pytest import approx
 
 pytestmark = pytest.mark.skipif(
-    "hybrid" not in su.solvers,
-    reason="hybrid not functional here"
+    "hybrid" not in su.solvers, reason="hybrid not functional here"
 )
 
 medium = load_qiime_medium(md.test_medium)
@@ -73,7 +72,7 @@ def test_community_objective(community):
     y = community.optimize(fluxes=True)
     assert isinstance(x, CommunitySolution)
     assert x.growth_rate == approx(0.873922, 1e-2, 1e-2)
-    assert x.members.growth_rate.dropna().sum() == approx(4*0.873922, 1e-2, 1e-2)
+    assert x.members.growth_rate.dropna().sum() == approx(4 * 0.873922, 1e-2, 1e-2)
     assert isinstance(y, CommunitySolution)
     assert y.fluxes.shape[0] == 5