+++ tests/test_problems.py
@@ -4,25 +4,24 @@ import numpy as np
 import pytest
 
 # Only test linear ones
-strategies = ['original', 'lmoma']
+strategies = ["original", "lmoma"]
 
 
-class TestOptcom():
+class TestOptcom:
 
     @pytest.mark.parametrize("strategy", strategies)
     def test_is_consistent(self, community, strategy):
         community.solver = "glpk"
         sol = community.optcom(strategy=strategy)
         assert np.allclose(sol.growth_rate, 0.873922)
-        assert np.allclose(sol.members.growth_rate.dropna().sum(), 4*0.873922)
+        assert np.allclose(sol.members.growth_rate.dropna().sum(), 4 * 0.873922)
 
-    @pytest.mark.parametrize("f", [0, 0.25, "0.5", [0, 0.2, 0.4, 0],
-                                   np.ones(4) * 0.7])
+    @pytest.mark.parametrize("f", [0, 0.25, "0.5", [0, 0.2, 0.4, 0], np.ones(4) * 0.7])
     def test_good_mingrowth(self, community, f):
         community.solver = "glpk"
         sol = community.optcom(strategy="lmoma", min_growth=f)
         assert np.allclose(sol.growth_rate, 0.873922)
-        assert np.allclose(sol.members.growth_rate.dropna().sum(), 4*0.873922)
+        assert np.allclose(sol.members.growth_rate.dropna().sum(), 4 * 0.873922)
 
     @pytest.mark.parametrize("f", ["a", [1, 2], np.ones(6)])
     def test_bad_mingrowth(self, community, f):