+++ tests/test_viz.py
@@ -8,8 +8,22 @@ import pytest
 import random
 
 
+def random_groups(res):
+    """Create some random groups for samples."""
+    groups = pd.Series(
+        random.choices(["a", "b"], k=4),
+        index=res.growth_rates.sample_id.unique(),
+        name="random",
+    )
+    return groups
+
+
 def test_plot_growth(growth_data, tmp_path):
     v = viz.plot_growth(growth_data, str(tmp_path / "viz.html"))
+    check_viz(v)
+    v = viz.plot_growth(
+        growth_data, str(tmp_path / "viz.html"), groups=random_groups(growth_data)
+    )
     check_viz(v)
 
 
@@ -22,48 +36,60 @@ def test_plot_sample_exchanges(growth_data, tmp_path):
     v = viz.plot_exchanges_per_sample(growth_data, str(tmp_path / "viz.html"))
     check_viz(v)
     v = viz.plot_exchanges_per_sample(
-        growth_data, str(tmp_path / "viz.html"), direction="export")
+        growth_data, str(tmp_path / "viz.html"), direction="export"
+    )
     check_viz(v)
     v = viz.plot_exchanges_per_sample(
-        growth_data, str(tmp_path / "viz.html"), cluster=False)
+        growth_data, str(tmp_path / "viz.html"), cluster=False
+    )
     check_viz(v)
     with pytest.raises(ValueError):
         v = viz.plot_exchanges_per_sample(
-            growth_data, str(tmp_path / "viz.html"), direction="dog")
+            growth_data, str(tmp_path / "viz.html"), direction="dog"
+        )
 
 
 def test_plot_taxon_exchanges(growth_data, tmp_path):
     v = viz.plot_exchanges_per_taxon(growth_data, str(tmp_path / "viz.html"))
     check_viz(v)
     v = viz.plot_exchanges_per_taxon(
-        growth_data, str(tmp_path / "viz.html"), direction="export")
+        growth_data, str(tmp_path / "viz.html"), direction="export"
+    )
+    check_viz(v)
+    v = viz.plot_exchanges_per_taxon(
+        growth_data, str(tmp_path / "viz.html"), groups=random_groups(growth_data)
+    )
     check_viz(v)
     with pytest.raises(ValueError):
         v = viz.plot_exchanges_per_taxon(
-            growth_data, str(tmp_path / "viz.html"), direction="dog")
+            growth_data, str(tmp_path / "viz.html"), direction="dog"
+        )
 
 
 def test_association(growth_data, tmp_path):
-    meta = pd.Series([0, 0, 1, 1],
-                     index=growth_data.growth_rates.sample_id.unique())
-    v = viz.plot_association(growth_data, meta, filename=str(tmp_path / "viz.html"),
-                     fdr_threshold=0.5)
+    meta = pd.Series([0, 0, 1, 1], index=growth_data.growth_rates.sample_id.unique())
+    v = viz.plot_association(
+        growth_data, meta, filename=str(tmp_path / "viz.html"), fdr_threshold=0.5
+    )
     check_viz(v)
-    v = viz.plot_association(growth_data, meta, variable_type="continuous",
-                     filename=str(tmp_path / "viz.html"), fdr_threshold=0.5)
+    v = viz.plot_association(
+        growth_data,
+        meta,
+        variable_type="continuous",
+        filename=str(tmp_path / "viz.html"),
+        fdr_threshold=0.5,
+    )
     check_viz(v)
 
     with pytest.raises(ValueError):
-        v = viz.plot_association(growth_data, meta, variable_type="dog",
-                         filename=str(tmp_path / "viz.html")
+        v = viz.plot_association(
+            growth_data, meta, variable_type="dog", filename=str(tmp_path / "viz.html")
         )
 
 
 def test_plot_focal_interactions(growth_data, tmp_path):
     v = viz.plot_focal_interactions(
-        growth_data,
-        taxon="Escherichia_coli_2",
-        filename=str(tmp_path / "viz.html")
+        growth_data, taxon="Escherichia_coli_2", filename=str(tmp_path / "viz.html")
     )
     check_viz(v)
 
@@ -71,10 +97,9 @@ def test_plot_focal_interactions(growth_data, tmp_path):
 def test_plot_mes(growth_data, tmp_path):
     v = viz.plot_mes(growth_data, filename=str(tmp_path / "viz.html"))
     check_viz(v)
-    groups = pd.Series(
-        random.choices(["a", "b"], k=4),
-        index=growth_data.growth_rates.sample_id.unique(),
-        name="random"
+    v = viz.plot_mes(
+        growth_data,
+        groups=random_groups(growth_data),
+        filename=str(tmp_path / "viz.html"),
     )
-    v = viz.plot_mes(growth_data, groups=groups, filename=str(tmp_path / "viz.html"))
     check_viz(v)